+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[2.1.0]
+- Delete video after downloading it from the camera
+- Add methods to get all files from the camera
+
 [2.0.1]
 - Make it work with Canon EOS M50